refactor: refactoring with hamin feddback ✌️
