+++ feature/mapMatching/src/main/java/com/mookiehare/hohoi/feature/mapmatching/MapMatchingScreen.kt
@@ -16,6 +16,7 @@ import androidx.compose.ui.platform.LocalView
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
+import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.google.accompanist.permissions.ExperimentalPermissionsApi
 import com.google.accompanist.permissions.PermissionsRequired
 import com.google.accompanist.permissions.rememberMultiplePermissionsState
@@ -25,10 +26,10 @@ import com.google.android.gms.maps.model.LatLng
 import com.google.maps.android.compose.*
 import com.mookiehare.hohoi.core.design.component.HelloiChip
 import com.mookiehare.hohoi.core.design.theme.HohoiTheme
-import com.mookiehare.hohoi.feature.mapmatching.model.RandomChip
-import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.launch
 import com.google.maps.android.clustering.ClusterManager
+import com.mookiehare.hohoi.core.design.component.RandomChip
+import com.mookiehare.hohoi.feature.mapmatching.model.Location
 import com.mookiehare.hohoi.feature.mapmatching.model.MarkerItem
 
 val chipsDummyData  =
@@ -50,16 +51,20 @@ internal fun MapMatchingRoute(
 ) {
 
     val items = remember { mutableStateListOf<MarkerItem>() }
+    items.clear()
     items.add(MarkerItem(LatLng(37.5480947960,126.87617507), "Marker", "Snippet"))
     items.add(MarkerItem(LatLng(37.548347622,126.87715100), "Marker", "Snippet"))
     items.add(MarkerItem(LatLng(37.548147622,126.87715100), "Marker", "Snippet"))
     items.add(MarkerItem(LatLng(37.548547622,126.87715100), "Marker", "Snippet"))
 //    items.add(MarkerItem(LatLng(37.548247622,126.87715100), "Marker", "Snippet"))
 
+    val locationState by viewModel.lastSelectedLocation.collectAsStateWithLifecycle()
+    
     MapMatchingScreen(
         modifier = modifier,
         viewModel,
-        items
+        items,
+        locationState
     )
 }
 
@@ -68,18 +73,18 @@ internal fun MapMatchingRoute(
 internal fun MapMatchingScreen(
     modifier: Modifier = Modifier,
     viewModel: MapMatchingViewModel = hiltViewModel(),
-    items: List<MarkerItem> = listOf()
+    items: List<MarkerItem> = listOf(),
+    location: Location
 ) {
 
-    val scope = rememberCoroutineScope()
     val activity = LocalView.current.context as Activity
     val multiplePermissionState = rememberMultiplePermissionsState(
         permissions = listOf(
-            Manifest.permission.ACCESS_COARSE_LOCATION
+            Manifest.permission.ACCESS_COARSE_LOCATION,
+            Manifest.permission.ACCESS_FINE_LOCATION
         )
     )
 
-    val locationState = MutableStateFlow(viewModel.getInitialLocation())
     val cameraPositionState = rememberCameraPositionState()
 
     LaunchedEffect(Unit) {
@@ -93,18 +98,15 @@ internal fun MapMatchingScreen(
     ){
         viewModel.getCurrentLocation(activity)
 
-        LaunchedEffect(viewModel.lastSelectedLocation) {
-            viewModel.lastSelectedLocation.collect {
-                locationState.value = it
-            }
-        }
-
-        scope.launch {
-            locationState.collect{
-                val loca = LatLng(it.lat, it.lng)
-                val cameraPosition = CameraPosition.fromLatLngZoom((loca), 15f)
-                cameraPositionState.animate(CameraUpdateFactory.newCameraPosition(cameraPosition), 1_000)
-            }
+        LaunchedEffect(location) {
+            cameraPositionState.animate(
+                CameraUpdateFactory.newCameraPosition(
+                    CameraPosition.fromLatLngZoom(
+                        LatLng(
+                            location.lat,
+                            location.lng
+                        ), 15f)),
+                1_000)
         }
 
         GoogleMap(
@@ -181,6 +183,6 @@ fun RandomMatchingChips(
 @Composable
 fun MapMatchingScreenPreview() {
     HohoiTheme {
-        MapMatchingScreen()
+//        MapMatchingScreen()
     }
 }
