--- feature/mapMatching/src/main/java/com/mookiehare/hohoi/feature/mapmatching/model/RandomChip.kt
@@ -1,8 +1,0 @@
-package com.mookiehare.hohoi.feature.mapmatching.model
-
-import androidx.compose.runtime.MutableState
-
-data class RandomChip (
-    val text : String = "",
-    val isSelected : MutableState<Boolean>
-    )
